+++ src/styles/style.scss
@@ -1,4 +1,3 @@
-@import '@fortawesome/fontawesome-svg-core/styles.css';
 body {
   --background-color: #343a40;
   --container-bg-color: #6c757d;
